LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/fetch/metadata/generated/element-picture.sub-expected.txt

Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY


FAIL sec-fetch-site - img[src] - Not sent to non-trustworthy same-origin destination, no attributes assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-site - img[srcset] - Not sent to non-trustworthy same-origin destination, no attributes assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-site - source[srcset] - Not sent to non-trustworthy same-origin destination, no attributes assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-site - img[src] - Not sent to non-trustworthy same-site destination, no attributes promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-site - img[srcset] - Not sent to non-trustworthy same-site destination, no attributes promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-site - source[srcset] - Not sent to non-trustworthy same-site destination, no attributes promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-site - img[src] - Not sent to non-trustworthy cross-site destination, no attributes assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-site - img[srcset] - Not sent to non-trustworthy cross-site destination, no attributes assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-site - source[srcset] - Not sent to non-trustworthy cross-site destination, no attributes assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-mode - img[src] - Not sent to non-trustworthy same-origin destination, no attributes assert_not_own_property: unexpected property "sec-fetch-mode" is found on object
FAIL sec-fetch-mode - img[srcset] - Not sent to non-trustworthy same-origin destination, no attributes assert_not_own_property: unexpected property "sec-fetch-mode" is found on object
FAIL sec-fetch-mode - source[srcset] - Not sent to non-trustworthy same-origin destination, no attributes assert_not_own_property: unexpected property "sec-fetch-mode" is found on object
FAIL sec-fetch-mode - img[src] - Not sent to non-trustworthy same-site destination, no attributes promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-mode - img[srcset] - Not sent to non-trustworthy same-site destination, no attributes promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-mode - source[srcset] - Not sent to non-trustworthy same-site destination, no attributes promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-mode - img[src] - Not sent to non-trustworthy cross-site destination, no attributes assert_not_own_property: unexpected property "sec-fetch-mode" is found on object
FAIL sec-fetch-mode - img[srcset] - Not sent to non-trustworthy cross-site destination, no attributes assert_not_own_property: unexpected property "sec-fetch-mode" is found on object
FAIL sec-fetch-mode - source[srcset] - Not sent to non-trustworthy cross-site destination, no attributes assert_not_own_property: unexpected property "sec-fetch-mode" is found on object
FAIL sec-fetch-dest - img[src] - Not sent to non-trustworthy same-origin destination, no attributes assert_not_own_property: unexpected property "sec-fetch-dest" is found on object
FAIL sec-fetch-dest - img[srcset] - Not sent to non-trustworthy same-origin destination, no attributes assert_not_own_property: unexpected property "sec-fetch-dest" is found on object
FAIL sec-fetch-dest - source[srcset] - Not sent to non-trustworthy same-origin destination, no attributes assert_not_own_property: unexpected property "sec-fetch-dest" is found on object
FAIL sec-fetch-dest - img[src] - Not sent to non-trustworthy same-site destination, no attributes promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-dest - img[srcset] - Not sent to non-trustworthy same-site destination, no attributes promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-dest - source[srcset] - Not sent to non-trustworthy same-site destination, no attributes promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-dest - img[src] - Not sent to non-trustworthy cross-site destination, no attributes assert_not_own_property: unexpected property "sec-fetch-dest" is found on object
FAIL sec-fetch-dest - img[srcset] - Not sent to non-trustworthy cross-site destination, no attributes assert_not_own_property: unexpected property "sec-fetch-dest" is found on object
FAIL sec-fetch-dest - source[srcset] - Not sent to non-trustworthy cross-site destination, no attributes assert_not_own_property: unexpected property "sec-fetch-dest" is found on object
PASS sec-fetch-user - img[src] - Not sent to non-trustworthy same-origin destination, no attributes
PASS sec-fetch-user - img[srcset] - Not sent to non-trustworthy same-origin destination, no attributes
PASS sec-fetch-user - source[srcset] - Not sent to non-trustworthy same-origin destination, no attributes
FAIL sec-fetch-user - img[src] - Not sent to non-trustworthy same-site destination, no attributes promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-user - img[srcset] - Not sent to non-trustworthy same-site destination, no attributes promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
FAIL sec-fetch-user - source[srcset] - Not sent to non-trustworthy same-site destination, no attributes promise_test: Unhandled rejection with value: object "Error: Failed to query for recorded headers."
PASS sec-fetch-user - img[src] - Not sent to non-trustworthy cross-site destination, no attributes
PASS sec-fetch-user - img[srcset] - Not sent to non-trustworthy cross-site destination, no attributes
PASS sec-fetch-user - source[srcset] - Not sent to non-trustworthy cross-site destination, no attributes
FAIL sec-fetch-site - img[src] - HTTPS downgrade (header not sent), no attributes assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-site - img[srcset] - HTTPS downgrade (header not sent), no attributes assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-site - source[srcset] - HTTPS downgrade (header not sent), no attributes assert_not_own_property: unexpected property "sec-fetch-site" is found on object
PASS sec-fetch-site - img[src] - HTTPS upgrade, no attributes
PASS sec-fetch-site - img[srcset] - HTTPS upgrade, no attributes
PASS sec-fetch-site - source[srcset] - HTTPS upgrade, no attributes
PASS sec-fetch-site - img[src] - HTTPS downgrade-upgrade, no attributes
PASS sec-fetch-site - img[srcset] - HTTPS downgrade-upgrade, no attributes
PASS sec-fetch-site - source[srcset] - HTTPS downgrade-upgrade, no attributes
